Allow RxJs 5.x
Remove dependency on `rx` and a load time check
for either `rx` or `rxjs`. This will also allow
consumers not using the SSE metrics observable
to exclude RxJs.

This also moves the test running to NPM, using
`npm run test-all` to test all the dependency configurations.
`npm test` and `gulp test` will now fail unless RxJs is installed.
Tests will fail if RxJs is added as a prod dependency, but not if
it is a dev or local dependency.