Metrics perf improvements
This change should improve the overall performance of HystrixJS by
leveraging some of the quirks of the V8 runtime. Specifically:

* Switching to `Map`s over dynamic objects, which is [significantly
  faster for property retrieval][perf-map]
* Using fixed-size Arrays for buckets in order to avoid the cost of
* Re-using buckets rather than garbage collecting and reinstantiating
  them.
* In `RollingPercentile`, re-using a global instance of `Stats`

In general, this will cause applications to pay the more memory cost of
running HystrixJS at startup, and increase less with volume, while also
reducing load on CPU and the garbage collector.

None of these changes alter the exposed interfaces.

[perf-map]: https://jsperf.com/javascript-objects-vs-map-performance
[perf-array]: https://jsperf.com/static-vs-dynamic-array/4